Update LESS references for renamed modular extensions.
