+++ console/src/components/Clock/Clock.module.less
@@ -1,4 +1,4 @@
-// Clock.less
+// Clock.module.less
 //
 
 .clock {
