+++ console/src/components/CompactMap/CompactMap.module.less
@@ -1,4 +1,4 @@
-// CompactMap.less
+// CompactMap.module.less
 //
 
 .compactMap {
