+++ console/src/components/CustomLayout/CustomLayout.module.less
@@ -1,4 +1,4 @@
-// CustomLayout.less
+// CustomLayout.module.less
 //
 
 @import "~@enact/agate/styles/mixins.less";
