+++ console/src/components/Dialer/Dialer.js
@@ -3,7 +3,7 @@ import kind from '@enact/core/kind';
 import {Row, Column, Cell} from '@enact/ui/Layout';
 import React from 'react';
 
-import css from './Dialer.less';
+import css from './Dialer.module.less';
 
 const Digit = kind({
 	name: 'Digit',
