+++ console/src/components/Dialer/Dialer.module.less
@@ -1,4 +1,4 @@
-// Dialer.less
+// Dialer.module.less
 //
 
 .dialer {
