+++ console/src/components/MapCore/MapCore.module.less
@@ -1,4 +1,4 @@
-// MapCore.less
+// MapCore.module.less
 //
 
 @import "https://api.tiles.mapbox.com/mapbox-gl-js/v0.49.0/mapbox-gl.css";
