+++ console/src/components/MapCore/Marker.module.less
@@ -1,4 +1,4 @@
-// Marker.less
+// Marker.module.less
 //
 
 @import "~@enact/agate/styles/skin.less";
