+++ console/src/components/ProfileDrawer/ProfileDrawer.module.less
@@ -1,4 +1,4 @@
-// ProfileDrawer.less
+// ProfileDrawer.module.less
 //
 @import "~@enact/agate/styles/skin.less";
