+++ console/src/components/UserAvatar/UserAvatar.module.less
@@ -1,4 +1,4 @@
-// UserAvatar.less
+// UserAvatar.module.less
 //
 @import "~@enact/agate/styles/mixins.less";
 @import "~@enact/agate/styles/skin.less";
