+++ console/src/components/UserSelectionPopup/UserSelectionPopup.module.less
@@ -1,4 +1,4 @@
-// UserSelectionPopup.less
+// UserSelectionPopup.module.less
 //
 @import "~@enact/agate/styles/skin.less";
