+++ console/src/components/WelcomePopup/WelcomePopup.module.less
@@ -1,4 +1,4 @@
-// WelcomePopup.less
+// WelcomePopup.module.less
 //
 
 @import "~@enact/agate/styles/skin.less";
