+++ console/src/views/AppList.js
@@ -7,7 +7,7 @@ import PropTypes from 'prop-types';
 import AppIconCell from '../components/AppIconCell';
 import {getPanelIndexOf} from '../App';
 
-import css from './AppList.less';
+import css from './AppList.module.less';
 
 const AppList = kind({
 	name: 'Home',
