+++ console/src/views/Home.module.less
@@ -1,4 +1,4 @@
-// Home.less
+// Home.module.less
 //
 
 @import "~@enact/agate/styles/mixins.less";
