+++ console/src/views/Map.module.less
@@ -1,4 +1,4 @@
-// Map.less
+// Map.module.less
 //
 
 .map {
