+++ sampler/src/AgateEnvironment/AgateEnvironment.module.less
@@ -1,4 +1,4 @@
-// // AgateEnvironment.less
+// // AgateEnvironment.module.less
 // //
 // @import "~@enact/moonstone/styles/variables.less";
 //
